Moved webpack file that was not in src anymore
