+++ lerna.json
@@ -3,5 +3,5 @@
     "packages/*",
     "examples/*"
   ],
-  "version": "0.0.0"
+  "version": "0.0.1-beta.0"
 }
