+++ packages/fabricsjs/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fabricsjs",
-  "version": "0.0.1",
+  "version": "0.0.1-alpha.0",
   "description": "Nextjs like framework for creating fragments for micro frontend applications",
   "main": "./src/index.js",
   "scripts": {
